🌈 style(repo): 应用 clang-format 并修复 cast/braces
- 对变更范围内的 C++ 文件应用 clang-format（SSOT：Qt6_CPP17_CLANG-FORMAT）
- 将示例/测试中的 C 风格 cast 替换为 static_cast，并补齐单语句控制流 braces
- 更新 .gitignore 忽略 helloagents_backup_*/
- 验证：ctest 39/39 passed；libcurl_consistency gate(all+ext) 87 passed

---

🌈 style(repo): apply clang-format and fix casts/braces

- Apply clang-format to changed C++ files (SSOT: Qt6_CPP17_CLANG-FORMAT)
- Replace C-style casts with static_cast and add braces for single-line control flow in examples/tests
- Update .gitignore to ignore helloagents_backup_*/
- Verified: ctest 39/39 passed; libcurl_consistency gate(all+ext) 87 passed

+++ examples/MiddlewareDemo/main.cpp
@@ -88,7 +88,11 @@ public:
         qDebug() << "失败:" << errorCount;
         if (totalRequests > 0) {
             qDebug() << "成功率:"
-                     << QString::number((double) successCount / totalRequests * 100, 'f', 2) << "%";
+                     << QString::number(static_cast<double>(successCount)
+                                            / static_cast<double>(totalRequests) * 100.0,
+                                        'f',
+                                        2)
+                     << "%";
         }
     }
 };
